implement petition creation

+++ src/main/java/org/example/Receiver/ReceiverRepository.java
@@ -2,5 +2,5 @@ package org.example.Receiver;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceiverRepository extends JpaRepository<Receiver, Long> {
+public interface ReceiverRepository extends JpaRepository<Receiver, Integer> {
 }
